// frontend/utils/types.ts
export enum Status {
  pending = "pending",
  paid = "paid",
  rejected = "rejected",
  review = "review",
  in_reposition = "in_reposition",
}

export type TableMode = "requests" | "reposiciones";
export type RequestType = "discount" | "expense";

export interface BaseEntity {
  id?: number | string;
  created_at?: string;
  updated_at?: string;
}

export interface RequestProps extends BaseEntity {
  unique_id: string;
  updated_at: string;
  invoice_number: number | string;
  account_id: string;
  account_name?: string;
  project_name?: string;
  amount: number;
  project: string | string[];
  responsible_id: string | null;
  vehicle_plate: string | null;
  vehicle_number: string | null;
  note: string | null;
  status: Status;
  type: RequestType;
  attachment_path: string | null;
  personnel_type?: string;
  request_date?: string;
  when?: string;
  month?: string;
  reposicion_id?: string;
}

export interface AccountProps extends BaseEntity {
  name: string;
  account_number?: string;
  account_type?: string;
  account_status?: string;
  account_affects?: string;
  generates_income?: boolean;
}

export interface ResponsibleProps {
  id: string | number;
  nombre_completo: string;
  proyecto: string;
}

export interface TransportProps {
  vehicle_plate: string;
  vehicle_number: string;
}